PublishVideo split into multiple components for cleaner code
