+++ src/components/common/Notification/Notification.tsx
@@ -1,6 +1,7 @@
 import { toast, ToastContainer, Zoom, Slide } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import { useAtom, useSetAtom } from 'jotai';
+import { useEffect } from 'react';
 import {
   alertAtom,
   removeNotificationAtom,
@@ -10,45 +11,47 @@ const Notification = () => {
   const [alertTypes] = useAtom(alertAtom);
   const removeNotification = useSetAtom(removeNotificationAtom);
 
-  if (alertTypes.alertError) {
-    toast.error(`❌ ${alertTypes?.alertError}`, {
-      position: 'bottom-right',
-      autoClose: 4000,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-      progress: undefined,
-      icon: false,
-    });
-    removeNotification();
-  }
-  if (alertTypes.alertSuccess) {
-    toast.success(`✔️ ${alertTypes?.alertSuccess}`, {
-      position: 'bottom-right',
-      autoClose: 4000,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-      progress: undefined,
-      icon: false,
-    });
-    removeNotification();
-  }
-  if (alertTypes.alertInfo) {
-    toast.info(`${alertTypes?.alertInfo}`, {
-      position: 'top-right',
-      autoClose: 1300,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-      progress: undefined,
-      theme: 'light',
-    });
-    removeNotification();
-  }
+  useEffect(() => {
+    if (alertTypes.alertError) {
+      toast.error(`❌ ${alertTypes?.alertError}`, {
+        position: 'bottom-right',
+        autoClose: 4000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        icon: false,
+      });
+      removeNotification();
+    }
+    if (alertTypes.alertSuccess) {
+      toast.success(`✔️ ${alertTypes?.alertSuccess}`, {
+        position: 'bottom-right',
+        autoClose: 4000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        icon: false,
+      });
+      removeNotification();
+    }
+    if (alertTypes.alertInfo) {
+      toast.info(`${alertTypes?.alertInfo}`, {
+        position: 'top-right',
+        autoClose: 1300,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+        theme: 'light',
+      });
+      removeNotification();
+    }
+  }, [alertTypes, removeNotification]);
 
   if (alertTypes.alertInfo) {
     return (
